+++ wsi_service_base_plugins/tifffile/wsi_service_plugin_tifffile/slide.py
@@ -215,7 +215,7 @@ class Slide(BaseSlide):
                     if fh.tell() != offset:
                         raise HTTPException(status_code=500, detail="Failed reading to tile offset")
                     data = fh.read(bytecount)
-                    tile, _indices, _shape = page.decode(data, index, jpegtables=jpegtables)
+                    tile, _, _ = page.decode(data, index, jpegtables=jpegtables)
 
                     # insert tile in temporary output array
                     tile_position_i = (i - start_tile_x0) * tile_height